+++ src/Lib/VivLib/Models/FceColor.cs
@@ -41,8 +41,8 @@ public record struct FceColor(int Hue, int Saturation, int Brightness, int Alpha
         int i = (int)Math.Floor(h);
         double f = h - i;
         double p = v * (1.0 - s);
-        double q = v * (1.0 - s * f);
-        double t = v * (1.0 - s * (1.0 - f));
+        double q = v * (1.0 - (s * f));
+        double t = v * (1.0 - (s * (1.0 - f)));
 
         return i switch
         {
